Make area page dynamic, merge globe component into parent file
