+++ src/app/area/[areaId]/page.tsx
@@ -1,7 +1,13 @@
-const AreaPage = () => {
+const AreaPage = ({
+  params,
+}: {
+  params: {
+    areaId: string;
+  };
+}) => {
   return (
     <div className="overflow-y-auto w-128 p-4 space-y-4">
-      This is area page!
+      This is area page for {params.areaId}
     </div>
   );
 };
